Add a script to compile a corpus
This patch adds a script that enables compiling a corpus with the
regalloc trace worker. This is intended for use when partially compiling
a corpus and enables extracting out only the functions that matter for
register allocation decisions.

This is intended to be used before training begins to run all the middle
end optimizations. This then allows for functions to be extracted after
all IPO has taken place so we can then compile them all separately
by just running the backend. This script also supports compiling all the
way to asm to enable testing with basic_block_trace_model.

This script is typically used with --mode=full so that basic blocks in
the trace that are in functions that do not have regalloc decisions can
get serialized. The script is then run again with --mode=bc to run all
middle end optimziations so that the extract_functions script can be run
on the partially compiled corpus to extract all of the functions with
regalloc decisions that we are interested in compiling during training.
--mode=asm is used for validating that an extracted corpus can be used
for modelling without having to start the training process.

Pull Request: https://github.com/google/ml-compiler-opt/pull/484