Add shuffle_repeat_count parameter to dataset function
This patch adds a shuffle_repeat_count parameter to tfrecord/file
datasets. This enables setting the repeat_count to something other
than None. Setting the repeat_count to None means it will repeat
indefinitely making it difficult to know when we have iterated over the
entire dataset. Setting it explicitly allows us to iterate over the
entire dataset a precise number of times when needed.